--- HigherSelf/app/data/HigherSelf_StorageKeys.ts
@@ -1,4 +1,0 @@
- export const STORAGE_KEYS = {
-    SAVED_AFFIRMATIONS: 'saved_affirmations',
-    DAILY_MOOD_SELECTIONS: 'daily_mood_selections',
-};
