+++ HigherSelf/data/affirmation.ts
@@ -1,0 +1,172 @@
+export const AFFIRMATIONS = [
+  { id: '1', text: 'I am disciplined and in control of my life.', category: 'discipline' },
+  { id: '2', text: 'I deserve peace, clarity, and confidence.', category: 'peace' },
+  { id: '3', text: 'I attract growth, focus, and abundance.', category: 'wealth' },
+
+  { id: '4', text: 'I am focused and committed to my goals.', category: 'discipline' },
+  { id: '5', text: 'I show up every day with purpose.', category: 'discipline' },
+  { id: '6', text: 'I stay consistent even when it is hard.', category: 'discipline' },
+  { id: '7', text: 'I take control of my habits and my future.', category: 'discipline' },
+  { id: '8', text: 'I finish what I start.', category: 'discipline' },
+  { id: '9', text: 'I am building a powerful and disciplined mind.', category: 'discipline' },
+  { id: '10', text: 'I follow through on my commitments.', category: 'discipline' },
+
+  { id: '11', text: 'I am calm, grounded, and at peace.', category: 'peace' },
+  { id: '12', text: 'I release stress and welcome clarity.', category: 'peace' },
+  { id: '13', text: 'I trust the process of my life.', category: 'peace' },
+  { id: '14', text: 'I breathe deeply and stay present.', category: 'peace' },
+  { id: '15', text: 'I let go of what I cannot control.', category: 'peace' },
+  { id: '16', text: 'I choose peace over worry.', category: 'peace' },
+
+  { id: '17', text: 'I am attracting financial abundance into my life.', category: 'wealth' },
+  { id: '18', text: 'I am open to new opportunities for success.', category: 'wealth' },
+  { id: '19', text: 'Money flows to me naturally and consistently.', category: 'wealth' },
+  { id: '20', text: 'I am building long-term wealth.', category: 'wealth' },
+  { id: '21', text: 'I deserve financial freedom.', category: 'wealth' },
+
+  { id: '22', text: 'I am confident in who I am.', category: 'confidence' },
+  { id: '23', text: 'I believe in my abilities.', category: 'confidence' },
+  { id: '24', text: 'I trust myself to make the right decisions.', category: 'confidence' },
+  { id: '25', text: 'I walk with confidence and purpose.', category: 'confidence' },
+  { id: '26', text: 'I am proud of the person I am becoming.', category: 'confidence' },
+
+  { id: '27', text: 'I grow stronger every single day.', category: 'growth' },
+  { id: '28', text: 'I am constantly evolving into my best self.', category: 'growth' },
+  { id: '29', text: 'I learn from every experience.', category: 'growth' },
+  { id: '30', text: 'I embrace growth and change.', category: 'growth' },
+
+  { id: '31', text: 'I stay disciplined even when motivation fades.', category: 'discipline' },
+  { id: '32', text: 'I am in control of my actions and reactions.', category: 'discipline' },
+  { id: '33', text: 'I create structure that supports my success.', category: 'discipline' },
+  { id: '34', text: 'I push through resistance and keep going.', category: 'discipline' },
+  { id: '35', text: 'I choose progress over perfection.', category: 'discipline' },
+
+  { id: '36', text: 'I allow myself to feel calm and relaxed.', category: 'peace' },
+  { id: '37', text: 'I am safe, grounded, and secure.', category: 'peace' },
+  { id: '38', text: 'I let go of tension in my body.', category: 'peace' },
+  { id: '39', text: 'I embrace stillness and clarity.', category: 'peace' },
+
+  { id: '40', text: 'I am aligned with success and abundance.', category: 'wealth' },
+  { id: '41', text: 'I create value that brings me wealth.', category: 'wealth' },
+  { id: '42', text: 'I am financially intelligent and responsible.', category: 'wealth' },
+
+  { id: '43', text: 'I am worthy of success and recognition.', category: 'confidence' },
+  { id: '44', text: 'I speak with clarity and confidence.', category: 'confidence' },
+  { id: '45', text: 'I trust my voice and my ideas.', category: 'confidence' },
+
+  { id: '46', text: 'I am becoming better every day.', category: 'growth' },
+  { id: '47', text: 'I welcome challenges that help me grow.', category: 'growth' },
+  { id: '48', text: 'I stay committed to my vision every day.', category: 'discipline' },
+  { id: '49', text: 'I trust that I am on the right path.', category: 'peace' },
+  { id: '50', text: 'I am creating the life I truly want.', category: 'growth' },
+  { id: '51', text: 'I show up for myself every single day.', category: 'discipline' },
+  { id: '52', text: 'I take full ownership of my actions.', category: 'discipline' },
+  { id: '53', text: 'I build habits that align with my goals.', category: 'discipline' },
+  { id: '54', text: 'I stay focused on what truly matters.', category: 'discipline' },
+  { id: '55', text: 'I follow through no matter how I feel.', category: 'discipline' },
+  { id: '56', text: 'I manage my time with intention and purpose.', category: 'discipline' },
+  { id: '57', text: 'I stay committed even when it gets uncomfortable.', category: 'discipline' },
+  { id: '58', text: 'I prioritize my growth over distractions.', category: 'discipline' },
+  { id: '59', text: 'I do what needs to be done without hesitation.', category: 'discipline' },
+  { id: '60', text: 'I am consistent in my pursuit of success.', category: 'discipline' },
+
+  { id: '61', text: 'I release all tension and embrace calm.', category: 'peace' },
+  { id: '62', text: 'I am grounded in the present moment.', category: 'peace' },
+  { id: '63', text: 'I choose calm over chaos.', category: 'peace' },
+  { id: '64', text: 'I create peace within myself.', category: 'peace' },
+  { id: '65', text: 'I am at ease with where I am right now.', category: 'peace' },
+  { id: '66', text: 'I trust that everything is working out for me.', category: 'peace' },
+  { id: '67', text: 'I allow myself to slow down and breathe.', category: 'peace' },
+  { id: '68', text: 'I welcome clarity into my mind.', category: 'peace' },
+  { id: '69', text: 'I am safe, calm, and supported.', category: 'peace' },
+  { id: '70', text: 'I let go of stress and embrace stillness.', category: 'peace' },
+
+  { id: '71', text: 'I attract wealth through my actions and mindset.', category: 'wealth' },
+  { id: '72', text: 'I am open to receiving financial success.', category: 'wealth' },
+  { id: '73', text: 'I create opportunities that bring me income.', category: 'wealth' },
+  { id: '74', text: 'I am aligned with financial growth.', category: 'wealth' },
+  { id: '75', text: 'I make smart and confident financial decisions.', category: 'wealth' },
+  { id: '76', text: 'I deserve to live a life of abundance.', category: 'wealth' },
+  { id: '77', text: 'I am building multiple streams of income.', category: 'wealth' },
+  { id: '78', text: 'I am worthy of financial success.', category: 'wealth' },
+  { id: '79', text: 'I attract profitable opportunities.', category: 'wealth' },
+  { id: '80', text: 'I manage money with confidence and clarity.', category: 'wealth' },
+
+  { id: '81', text: 'I trust myself completely.', category: 'confidence' },
+  { id: '82', text: 'I am comfortable being myself.', category: 'confidence' },
+  { id: '83', text: 'I carry myself with confidence and strength.', category: 'confidence' },
+  { id: '84', text: 'I believe in my vision and my path.', category: 'confidence' },
+  { id: '85', text: 'I am proud of who I am becoming.', category: 'confidence' },
+  { id: '86', text: 'I speak and act with certainty.', category: 'confidence' },
+  { id: '87', text: 'I deserve respect and success.', category: 'confidence' },
+  { id: '88', text: 'I am bold in pursuing my goals.', category: 'confidence' },
+  { id: '89', text: 'I trust my instincts and decisions.', category: 'confidence' },
+  { id: '90', text: 'I am confident in every situation I face.', category: 'confidence' },
+
+  { id: '91', text: 'I grow stronger with every challenge.', category: 'growth' },
+  { id: '92', text: 'I am committed to becoming better daily.', category: 'growth' },
+  { id: '93', text: 'I embrace learning and improvement.', category: 'growth' },
+  { id: '94', text: 'I evolve into a better version of myself.', category: 'growth' },
+  { id: '95', text: 'I welcome new opportunities to grow.', category: 'growth' },
+  { id: '96', text: 'I improve my mindset every day.', category: 'growth' },
+  { id: '97', text: 'I am always progressing forward.', category: 'growth' },
+  { id: '98', text: 'I learn from every experience I face.', category: 'growth' },
+  { id: '99', text: 'I am becoming stronger, smarter, and wiser.', category: 'growth' },
+  { id: '100', text: 'I grow through discipline and consistency.', category: 'growth' },
+  { id: '101', text: 'I take action even when I feel uncertain.', category: 'discipline' },
+  { id: '102', text: 'I stay consistent with the habits that serve me.', category: 'discipline' },
+  { id: '103', text: 'I honor my commitments to myself.', category: 'discipline' },
+  { id: '104', text: 'I focus on what I can control and act on it.', category: 'discipline' },
+  { id: '105', text: 'I follow a routine that supports my success.', category: 'discipline' },
+  { id: '106', text: 'I move forward with clarity and intention.', category: 'discipline' },
+  { id: '107', text: 'I choose discipline over distraction.', category: 'discipline' },
+  { id: '108', text: 'I complete important tasks before comfort.', category: 'discipline' },
+  { id: '109', text: 'I show up with discipline every single day.', category: 'discipline' },
+  { id: '110', text: 'I build momentum through small daily wins.', category: 'discipline' },
+
+  { id: '111', text: 'I return to calm with every breath I take.', category: 'peace' },
+  { id: '112', text: 'I allow peace to guide my thoughts.', category: 'peace' },
+  { id: '113', text: 'I am centered and balanced in all situations.', category: 'peace' },
+  { id: '114', text: 'I choose stillness in moments of stress.', category: 'peace' },
+  { id: '115', text: 'I am at peace with my journey.', category: 'peace' },
+  { id: '116', text: 'I let go of worry and embrace trust.', category: 'peace' },
+  { id: '117', text: 'I create a calm space within my mind.', category: 'peace' },
+  { id: '118', text: 'I respond with calm and clarity.', category: 'peace' },
+  { id: '119', text: 'I feel grounded, safe, and secure.', category: 'peace' },
+  { id: '120', text: 'I carry peace with me wherever I go.', category: 'peace' },
+
+  { id: '121', text: 'I welcome financial growth into my life.', category: 'wealth' },
+  { id: '122', text: 'I am capable of building lasting wealth.', category: 'wealth' },
+  { id: '123', text: 'I take smart steps toward financial freedom.', category: 'wealth' },
+  { id: '124', text: 'I attract clients and opportunities with ease.', category: 'wealth' },
+  { id: '125', text: 'I increase my income through value and effort.', category: 'wealth' },
+  { id: '126', text: 'I am aligned with profitable decisions.', category: 'wealth' },
+  { id: '127', text: 'I use money as a tool to build my future.', category: 'wealth' },
+  { id: '128', text: 'I am confident in my financial journey.', category: 'wealth' },
+  { id: '129', text: 'I grow my wealth through consistency.', category: 'wealth' },
+  { id: '130', text: 'I am open to receiving unexpected income.', category: 'wealth' },
+
+  { id: '131', text: 'I believe in my ability to succeed.', category: 'confidence' },
+  { id: '132', text: 'I carry confidence into every room I enter.', category: 'confidence' },
+  { id: '133', text: 'I am comfortable taking up space.', category: 'confidence' },
+  { id: '134', text: 'I express myself clearly and confidently.', category: 'confidence' },
+  { id: '135', text: 'I am worthy of opportunities and success.', category: 'confidence' },
+  { id: '136', text: 'I handle challenges with confidence.', category: 'confidence' },
+  { id: '137', text: 'I trust my preparation and my ability.', category: 'confidence' },
+  { id: '138', text: 'I step outside my comfort zone with courage.', category: 'confidence' },
+  { id: '139', text: 'I am secure in who I am.', category: 'confidence' },
+  { id: '140', text: 'I radiate confidence and strength.', category: 'confidence' },
+
+  { id: '141', text: 'I embrace change as a path to growth.', category: 'growth' },
+  { id: '142', text: 'I am always learning and improving.', category: 'growth' },
+  { id: '143', text: 'I turn obstacles into opportunities.', category: 'growth' },
+  { id: '144', text: 'I evolve with every experience I have.', category: 'growth' },
+  { id: '145', text: 'I commit to becoming my best self.', category: 'growth' },
+  { id: '146', text: 'I grow stronger through discipline.', category: 'growth' },
+  { id: '147', text: 'I expand my mindset every day.', category: 'growth' },
+  { id: '148', text: 'I welcome feedback and use it to improve.', category: 'growth' },
+  { id: '149', text: 'I take steps forward, no matter how small.', category: 'growth' },
+  { id: '150', text: 'I am dedicated to lifelong growth.', category: 'growth' },
+];
+
+export default AFFIRMATIONS;
