+++ HigherSelf/types/affirmations.ts
